requirements.txt
dash==2.14.1
plotly==5.17.0
pandas==2.1.1
numpy==1.24.3
duckdb==0.9.1
pytest==7.4.2
python-dotenv==1.0.0